feat: Montyhall Done
Suppose you're on a game show and you're given the choice of three doors.
Behind one door is a car; behind the others, goats.
The car and the goats were placed randomly behind the doors before the
show.
Run random simulations of the Monty Hall game. Show the effects of a
strategy of the contestant always keeping his first guess so it can be contrasted
with the strategy of the contestant always switching his guess.
Simulate at least a thousand games using three doors for each strategy and
show the results in such a way as to make it easy to compare the effects of
each strategy.